requirements.txt
aiosqlite==0.17.0
alembic==1.7.7
altair==5.3.0
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.26.0
atomicwrites==1.4.1
attrs==23.2.0
bcrypt==3.2.2
blinker==1.8.2
cachetools==5.4.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==6.5.0
cryptography==37.0.2
dependency-injector==4.39.1
ecdsa==0.17.0
execnet==1.9.0
fastapi==0.93.0
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==1.1.2
h11==0.13.0
httpcore==0.16.3
httptools==0.4.0
httpx==0.23.3
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
isort==5.10.1
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
loguru==0.6.0
Mako==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
numpy==2.0.1
packaging==21.3
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
pluggy==1.0.0
protobuf==5.27.3
psycopg2-binary==2.9.4
py==1.11.0
pyarrow==17.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pydeck==0.9.1
pygments==2.18.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-cov==4.0.0
pytest-forked==1.4.0
pytest-mock==3.9.0
pytest-xdist==2.5.0
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
PyYAML==6.0
referencing==0.35.1
requests==2.28.1
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.8
six==1.16.0
smmap==5.0.1
sniffio==1.2.0
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
starlette==0.25.0
streamlit==1.37.0
tenacity==8.5.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==1.26.12
uvicorn==0.20.0
watchdog==4.0.1
watchfiles==0.17.0
watchgod==0.8.2
websocket==0.2.1
websockets==10.3
win32-setctime==1.1.0
zipp==3.8.0
zope.event==5.0
zope.interface==6.4.post2
